fix: application id, 공고 상세 정보 로직 커스텀 훅 분리
